add a test button.
1. For a local service without <android:process=":process_name"> specified, it will run in the main thread of its hosting process, i.e. the UI thread of the application’s main process.
So methods like onCreate, onStart, onDestroy are all run in the UI thread. Therefore, we can propose a Toast in these methods.
It should spawn its own thread in which to do time-consuming work, and ParaTask can do that job. In this situation, ParaTask’s slot handling thread is just the UI thread of the app.

2. For a local service specified <android:process=":process_name">, it will run in its own process. That process has its own UI thread with a running Looper, so a Toast proposed in the service’s onCreate or onStart methods won’t cause an exception as follows:

java.lang.RuntimeException: Can't create handler inside thread that has not called Looper.prepare()

If we use ParaTask in such a service, the slot handling thread will be the UI thread of the service process, and we can propose toasts in that UI thread.

But why does the toast look like running in the app’s main thread? Yes, just looks like. Not really. A toast can pop up at the desktop, even if there is no Activity running at foreground. So toast is a special kind of UI widget, which can show up on top of all activities.